+++ tybotRoute/CHANGELOG.md
@@ -5,6 +5,19 @@
 available on:
  ▶️ https://www.npmjs.com/package/@tiledesk/tiledesk-tybot-connector
 
+# v0.2.58 (next-release)
+- Added support for DTMF Menu (Reply for voice channel)
+
+# v0.2.58-rc1
+- Added support for DTMF Menu (Reply for voice channel)
+
+# v0.2.57
+- Changed env variables for AskGPT actions
+- Added support for temperature, top_k, max_tokens and context for AskGPTV2 action
+- Added support for context with variables for GptTask action
+- fixed catching reply error in index.js: reply = await chatbot.replyToMessage(message);
+- Added support for DTMF Form and Blind Transfer (Reply for voice channel)
+
 # v0.2.56
 - Fixing Customerio Action not passing testing